requirement.txt
openai
python-dotenv
gradio
streamlit
google-genai
google-generativeai

